split HmacSpec into core microcrate
