+++ crates/uselesskey-core-hmac-spec/README.md
@@ -1,0 +1,6 @@
+# uselesskey-core-hmac-spec
+
+Core `HmacSpec` model for uselesskey's test-fixture crates.
+
+This crate contains the stable algorithm enum and deterministic cache-key encoding
+used by higher-level HMAC fixture crates.
